embedded/raspberry-pi/conveyer_belt/control_app/requirements.txt
rpi-lgpio
paho-mqtt
smbus
